Feature/fix toolbar button (#52)
