+++ app/app.vue
@@ -7,6 +7,6 @@
 </template>
 
 <script setup lang="ts">
-import { SpeedInsights } from '@vercel/speed-insights/vue'
-import { Analytics } from '@vercel/analytics/nuxt'
+  import { SpeedInsights } from '@vercel/speed-insights/vue'
+  import { Analytics } from '@vercel/analytics/nuxt'
 </script>
