+++ app/components/LatexExample.vue
@@ -3,100 +3,95 @@
 </template>
 
 <script setup lang="ts">
-import { ref, onMounted, watch } from 'vue'
-import { useMarkdownRenderer } from '../composables/useMarkdownRenderer'
+  import { ref, onMounted, watch } from 'vue'
+  import { useMarkdownRenderer } from '../composables/useMarkdownRenderer'
 
-interface Props {
-  latex: string
-  isBlock?: boolean
-}
+  interface Props {
+    latex: string
+    isBlock?: boolean
+  }
 
-const props = withDefaults(defineProps<Props>(), {
-  isBlock: false,
-})
+  const props = withDefaults(defineProps<Props>(), {
+    isBlock: false,
+  })
 
-const { renderLatexExample } = useMarkdownRenderer()
-const renderedContent = ref('')
+  const { renderLatexExample } = useMarkdownRenderer()
+  const renderedContent = ref('')
 
-// Wait for KaTeX script to be available on the client
-const waitForKatex = (): Promise<void> => {
-  return new Promise((resolve) => {
-    if (!import.meta.client) {
-      resolve()
-      return
-    }
-    if ((window as any).katex) {
-      resolve()
-      return
-    }
-    const check = () => {
+  // Wait for KaTeX script to be available on the client
+  const waitForKatex = (): Promise<void> => {
+    return new Promise((resolve) => {
+      if (!import.meta.client) {
+        resolve()
+        return
+      }
       if ((window as any).katex) {
         resolve()
-      } else {
-        setTimeout(check, 50)
+        return
       }
-    }
-    check()
-  })
-}
-
-const renderContent = () => {
-  const input = props.latex || ''
-  if (!input) {
-    renderedContent.value = ''
-    return
+      const check = () => {
+        if ((window as any).katex) {
+          resolve()
+        } else {
+          setTimeout(check, 50)
+        }
+      }
+      check()
+    })
   }
 
-  try {
-    // If isBlock is true and input contains $$ blocks, split and render each block separately
-    if (props.isBlock) {
-      // Split by $$ while keeping delimiters to determine blocks
-      const parts = input.split(/(\$\$[\s\S]*?\$\$)/g).filter(Boolean)
-
-      const html = parts
-        .map((part) => {
-          const trimmed = part.trim()
-          if (/^\$\$[\s\S]*?\$\$$/.test(trimmed)) {
-            const content = trimmed.slice(2, -2) // remove $$ ... $$
-            return renderLatexExample(content, true)
-          }
-          // Non-math text, render as paragraph
-          return trimmed ? `<p>${trimmed.replace(/\n+/g, '<br/>')}</p>` : ''
-        })
-        .join('')
-
-      renderedContent.value = html
+  const renderContent = () => {
+    const input = props.latex || ''
+    if (!input) {
+      renderedContent.value = ''
       return
     }
 
-    // Inline mode: simply render the string (renderLatexExample already supports inline math symbols)
-    renderedContent.value = renderLatexExample(input, false)
-  } catch {
-    renderedContent.value = input
-  }
-}
+    try {
+      if (props.isBlock) {
+        const parts = input.split(/(\$\$[\s\S]*?\$\$)/g).filter(Boolean)
 
-watch(
-  () => [props.latex, props.isBlock],
-  async () => {
-    if (import.meta.client && !(window as any).katex) {
-      await waitForKatex()
+        const html = parts
+          .map((part) => {
+            const trimmed = part.trim()
+            if (/^\$\$[\s\S]*?\$\$$/.test(trimmed)) {
+              const content = trimmed.slice(2, -2)
+              return renderLatexExample(content, true)
+            }
+            return trimmed ? `<p>${trimmed.replace(/\n+/g, '<br/>')}</p>` : ''
+          })
+          .join('')
+
+        renderedContent.value = html
+        return
+      }
+
+      renderedContent.value = renderLatexExample(input, false)
+    } catch {
+      renderedContent.value = input
     }
-    renderContent()
-  },
-  { immediate: true }
-)
-
-onMounted(() => {
-  if (import.meta.client) {
-    // Ensure KaTeX is loaded before first render
-    waitForKatex().then(renderContent)
   }
-})
+
+  watch(
+    () => [props.latex, props.isBlock],
+    async () => {
+      if (import.meta.client && !(window as any).katex) {
+        await waitForKatex()
+      }
+      renderContent()
+    },
+    { immediate: true }
+  )
+
+  onMounted(() => {
+    if (import.meta.client) {
+      waitForKatex().then(renderContent)
+    }
+  })
 </script>
 
 <style scoped>
-.math-display {
-  line-height: 1.5;
-}
+  .math-display {
+    line-height: 1.5;
+  }
 </style>
