+++ app/layouts/default.vue
@@ -5,8 +5,8 @@
 </template>
 
 <script setup>
-// Add dark class to html element
-if (import.meta.client) {
-  document.documentElement.classList.add('dark')
-}
+  // Add dark class to html element
+  if (import.meta.client) {
+    document.documentElement.classList.add('dark')
+  }
 </script>
