+++ app/pages/editor.vue
@@ -138,399 +138,401 @@
 </template>
 
 <script setup lang="ts">
-import { ref, watch, computed, nextTick, onMounted, onUnmounted } from 'vue'
-import type { Ref } from 'vue'
-import { EditorView, keymap } from '@codemirror/view'
-import { basicSetup } from 'codemirror'
-import { EditorState } from '@codemirror/state'
-import { indentWithTab } from '@codemirror/commands'
-import { markdown } from '@codemirror/lang-markdown'
-import { oneDark } from '@codemirror/theme-one-dark'
+  import { ref, watch, computed, nextTick, onMounted, onUnmounted } from 'vue'
+  import type { Ref } from 'vue'
+  import { EditorView, keymap } from '@codemirror/view'
+  import { basicSetup } from 'codemirror'
+  import { EditorState } from '@codemirror/state'
+  import { indentWithTab } from '@codemirror/commands'
+  import { markdown } from '@codemirror/lang-markdown'
+  import { oneDark } from '@codemirror/theme-one-dark'
 
-import EditorToolbar from '../components/EditorToolbar.vue'
-import HelpModal from '../components/HelpModal.vue'
-import ConfirmModal from '../components/ConfirmModal.vue'
-import LoadingScreen from '../components/LoadingScreen.vue'
-import { useEditorActions } from '../composables/useEditorActions'
-import { useKeyboardShortcuts } from '../composables/useKeyboardShortcuts'
-import { useMarkdownRenderer } from '../composables/useMarkdownRenderer'
-import { PhQuestion } from '@phosphor-icons/vue'
+  import EditorToolbar from '../components/EditorToolbar.vue'
+  import HelpModal from '../components/HelpModal.vue'
+  import ConfirmModal from '../components/ConfirmModal.vue'
+  import LoadingScreen from '../components/LoadingScreen.vue'
+  import { useEditorActions } from '../composables/useEditorActions'
+  import { useKeyboardShortcuts } from '../composables/useKeyboardShortcuts'
+  import { useMarkdownRenderer } from '../composables/useMarkdownRenderer'
+  import { PhQuestion } from '@phosphor-icons/vue'
 
-// Add page-specific styling to prevent scrolling
-useHead({
-  bodyAttrs: {
-    class: 'overflow-hidden',
-  },
-})
+  // Add page-specific styling to prevent scrolling
+  useHead({
+    bodyAttrs: {
+      class: 'overflow-hidden',
+    },
+  })
 
-const isLoading: Ref<boolean> = ref(true)
-const editorContainer: Ref<HTMLElement | null> = ref(null)
-const previewContainer: Ref<HTMLElement | null> = ref(null)
-const content: Ref<string> = ref('')
-const showEditor: Ref<boolean> = ref(true)
-const showPreview: Ref<boolean> = ref(true)
-const autosave: Ref<boolean> = ref(false)
-const lastSaved: Ref<string> = ref('')
-const isMobile: Ref<boolean> = ref(false)
-const showHelpModal: Ref<boolean> = ref(false)
-const pendingAutosaveValue: Ref<boolean | null> = ref(null)
-const confirmAutosaveOn: Ref<boolean> = ref(false)
-const confirmAutosaveOff: Ref<boolean> = ref(false)
-const confirmClearData: Ref<boolean> = ref(false)
+  const isLoading: Ref<boolean> = ref(true)
+  const editorContainer: Ref<HTMLElement | null> = ref(null)
+  const previewContainer: Ref<HTMLElement | null> = ref(null)
+  const content: Ref<string> = ref('')
+  const showEditor: Ref<boolean> = ref(true)
+  const showPreview: Ref<boolean> = ref(true)
+  const autosave: Ref<boolean> = ref(false)
+  const lastSaved: Ref<string> = ref('')
+  const isMobile: Ref<boolean> = ref(false)
+  const showHelpModal: Ref<boolean> = ref(false)
+  const pendingAutosaveValue: Ref<boolean | null> = ref(null)
+  const confirmAutosaveOn: Ref<boolean> = ref(false)
+  const confirmAutosaveOff: Ref<boolean> = ref(false)
+  const confirmClearData: Ref<boolean> = ref(false)
 
-const editorViewRef: Ref<EditorView | null> = ref(null)
-const customEditorWidth: Ref<number | null> = ref(null)
-const customPreviewWidth: Ref<number | null> = ref(null)
+  const editorViewRef: Ref<EditorView | null> = ref(null)
+  const customEditorWidth: Ref<number | null> = ref(null)
+  const customPreviewWidth: Ref<number | null> = ref(null)
 
-const wordCount = computed<number>(() => {
-  return content.value.trim() ? content.value.trim().split(/\s+/).length : 0
-})
+  const wordCount = computed<number>(() => {
+    return content.value.trim() ? content.value.trim().split(/\s+/).length : 0
+  })
 
-const charCount = computed<number>(() => {
-  return content.value.length
-})
+  const charCount = computed<number>(() => {
+    return content.value.length
+  })
 
-const renderedContent: Ref<string> = ref('')
+  const renderedContent: Ref<string> = ref('')
 
-const { renderMarkdown, renderMermaidDiagrams, highlightSyntax } = useMarkdownRenderer()
-const actions = useEditorActions(editorViewRef, content, autosave, lastSaved)
-useKeyboardShortcuts(actions)
+  const { renderMarkdown, renderMermaidDiagrams, highlightSyntax } = useMarkdownRenderer()
+  const actions = useEditorActions(editorViewRef, content, autosave, lastSaved)
+  useKeyboardShortcuts(actions)
 
-// Watch content changes and render markdown
-watch(
-  content,
-  async (newContent) => {
-    renderedContent.value = await renderMarkdown(newContent)
-    await nextTick()
-    await renderMermaidDiagrams()
-    highlightSyntax()
-  },
-  { immediate: true }
-)
+  // Watch content changes and render markdown
+  watch(
+    content,
+    async (newContent) => {
+      renderedContent.value = await renderMarkdown(newContent)
+      await nextTick()
+      await renderMermaidDiagrams()
+      highlightSyntax()
+    },
+    { immediate: true }
+  )
 
-// Methods
-const onLoadingComplete = (): void => {
-  isLoading.value = false
-}
+  // Methods
+  const onLoadingComplete = (): void => {
+    isLoading.value = false
+  }
 
-const initEditor = (): void => {
-  if (!editorContainer.value) return
+  const initEditor = (): void => {
+    if (!editorContainer.value) return
 
-  const extensions = [
-    basicSetup,
-    markdown(),
-    oneDark,
-    EditorView.lineWrapping, 
-    keymap.of([indentWithTab]),
-    EditorView.updateListener.of((update) => {
-      if (update.docChanged) {
-        content.value = update.state.doc.toString()
+    const extensions = [
+      basicSetup,
+      markdown(),
+      oneDark,
+      EditorView.lineWrapping,
+      keymap.of([indentWithTab]),
+      EditorView.updateListener.of((update) => {
+        if (update.docChanged) {
+          content.value = update.state.doc.toString()
+        }
+      }),
+      EditorView.theme({
+        '&': {
+          height: '100%',
+          fontSize: '14px',
+        },
+        '.cm-editor': {
+          height: '100%',
+        },
+        '.cm-scroller': {
+          height: '100%',
+        },
+        '.cm-content': {
+          minHeight: '100%',
+          padding: '16px',
+        },
+      }),
+    ]
+
+    const state = EditorState.create({
+      doc: content.value,
+      extensions,
+    })
+
+    editorViewRef.value = new EditorView({
+      state,
+      parent: editorContainer.value,
+    })
+  }
+
+  const togglePreview = (): void => {
+    showPreview.value = !showPreview.value
+    if (!showPreview.value && !showEditor.value) {
+      showEditor.value = true
+    }
+    nextTick(() => applyPaneWidths())
+  }
+
+  const toggleEditor = (): void => {
+    showEditor.value = !showEditor.value
+    if (!showEditor.value && !showPreview.value) {
+      showPreview.value = true
+    }
+    nextTick(() => applyPaneWidths())
+  }
+
+  const applyPaneWidths = (): void => {
+    const editorPane = document.querySelector('.editor-pane') as HTMLElement
+    const previewPane = document.querySelector('.preview-pane') as HTMLElement
+
+    if (
+      showEditor.value &&
+      showPreview.value &&
+      customEditorWidth.value !== null &&
+      customPreviewWidth.value !== null
+    ) {
+      if (editorPane) {
+        editorPane.style.width = `${customEditorWidth.value}%`
       }
-    }),
-    EditorView.theme({
-      '&': {
-        height: '100%',
-        fontSize: '14px',
-      },
-      '.cm-editor': {
-        height: '100%',
-      },
-      '.cm-scroller': {
-        height: '100%',
-      },
-      '.cm-content': {
-        minHeight: '100%',
-        padding: '16px',
-      },
-    }),
-  ]
+      if (previewPane) {
+        previewPane.style.width = `${customPreviewWidth.value}%`
+      }
+    } else {
+      if (editorPane) {
+        editorPane.style.width = ''
+      }
+      if (previewPane) {
+        previewPane.style.width = ''
+      }
+    }
+  }
 
-  const state = EditorState.create({
-    doc: content.value,
-    extensions,
+  const startResize = (e: MouseEvent): void => {
+    e.preventDefault()
+
+    const startX = e.clientX
+    const editorPane = document.querySelector('.editor-pane') as HTMLElement
+    const previewPane = document.querySelector('.preview-pane') as HTMLElement
+
+    if (!editorPane || !previewPane) return
+
+    const startEditorWidth = editorPane.offsetWidth
+    const containerWidth = editorPane.parentElement!.offsetWidth
+
+    const handleMouseMove = (e: MouseEvent): void => {
+      const deltaX = e.clientX - startX
+      const newEditorWidth = startEditorWidth + deltaX
+      const newEditorPercent = (newEditorWidth / containerWidth) * 100
+
+      if (newEditorPercent >= 20 && newEditorPercent <= 80) {
+        editorPane.style.width = `${newEditorPercent}%`
+        previewPane.style.width = `${100 - newEditorPercent}%`
+      }
+    }
+
+    const handleMouseUp = (): void => {
+      document.removeEventListener('mousemove', handleMouseMove)
+      document.removeEventListener('mouseup', handleMouseUp)
+
+      const editorPercent = (editorPane.offsetWidth / containerWidth) * 100
+      const previewPercent = 100 - editorPercent
+
+      customEditorWidth.value = editorPercent
+      customPreviewWidth.value = previewPercent
+      savePaneWidths()
+    }
+
+    document.addEventListener('mousemove', handleMouseMove)
+    document.addEventListener('mouseup', handleMouseUp)
+  }
+
+  const setupScrollSync = (): void => {
+    if (!editorViewRef.value || !previewContainer.value) return
+
+    let isEditorScrolling = false
+    let isPreviewScrolling = false
+
+    const editorScrollHandler = (): void => {
+      if (isPreviewScrolling) return
+
+      isEditorScrolling = true
+
+      const editorElement = editorViewRef.value!.scrollDOM
+      const scrollPercentage =
+        editorElement.scrollTop / (editorElement.scrollHeight - editorElement.clientHeight)
+
+      const previewElement = previewContainer.value!
+      const maxScrollTop = previewElement.scrollHeight - previewElement.clientHeight
+
+      previewElement.scrollTop = scrollPercentage * maxScrollTop
+
+      setTimeout(() => {
+        isEditorScrolling = false
+      }, 50)
+    }
+
+    const previewScrollHandler = (): void => {
+      if (isEditorScrolling) return
+
+      isPreviewScrolling = true
+
+      const previewElement = previewContainer.value!
+      const scrollPercentage =
+        previewElement.scrollTop / (previewElement.scrollHeight - previewElement.clientHeight)
+
+      const editorElement = editorViewRef.value!.scrollDOM
+      const maxScrollTop = editorElement.scrollHeight - editorElement.clientHeight
+
+      editorElement.scrollTop = scrollPercentage * maxScrollTop
+
+      setTimeout(() => {
+        isPreviewScrolling = false
+      }, 50)
+    }
+
+    editorViewRef.value.scrollDOM.addEventListener('scroll', editorScrollHandler)
+    previewContainer.value.addEventListener('scroll', previewScrollHandler)
+  }
+
+  const loadSettings = (): void => {
+    try {
+      const savedAutosave = localStorage.getItem('mermditor-autosave')
+      if (savedAutosave !== null) {
+        autosave.value = savedAutosave === 'true'
+      }
+    } catch (error) {
+      console.error('Error loading settings:', error)
+    }
+  }
+
+  const loadContent = (): void => {
+    const saved = localStorage.getItem('mermditor-content')
+    if (saved) {
+      content.value = saved
+    }
+  }
+
+  const savePaneWidths = (): void => {
+    try {
+      if (customEditorWidth.value !== null && customPreviewWidth.value !== null) {
+        localStorage.setItem('mermditor-editor-width', customEditorWidth.value.toString())
+        localStorage.setItem('mermditor-preview-width', customPreviewWidth.value.toString())
+      }
+    } catch (error) {
+      console.error('Error saving pane widths:', error)
+    }
+  }
+
+  const loadPaneWidths = (): void => {
+    try {
+      const savedEditorWidth = localStorage.getItem('mermditor-editor-width')
+      const savedPreviewWidth = localStorage.getItem('mermditor-preview-width')
+
+      if (savedEditorWidth && savedPreviewWidth) {
+        customEditorWidth.value = parseFloat(savedEditorWidth)
+        customPreviewWidth.value = parseFloat(savedPreviewWidth)
+      }
+    } catch (error) {
+      console.error('Error loading pane widths:', error)
+    }
+  }
+
+  // Watchers
+  let saveTimeout: ReturnType<typeof setTimeout> | null = null
+
+  watch(content, () => {
+    if (autosave.value) {
+      if (saveTimeout) {
+        clearTimeout(saveTimeout)
+      }
+      saveTimeout = setTimeout(() => {
+        actions.saveContent()
+      }, 10000)
+    }
   })
 
-  editorViewRef.value = new EditorView({
-    state,
-    parent: editorContainer.value,
-  })
-}
-
-const togglePreview = (): void => {
-  showPreview.value = !showPreview.value
-  if (!showPreview.value && !showEditor.value) {
-    showEditor.value = true
-  }
-  nextTick(() => applyPaneWidths())
-}
-
-const toggleEditor = (): void => {
-  showEditor.value = !showEditor.value
-  if (!showEditor.value && !showPreview.value) {
-    showPreview.value = true
-  }
-  nextTick(() => applyPaneWidths())
-}
-
-const applyPaneWidths = (): void => {
-  const editorPane = document.querySelector('.editor-pane') as HTMLElement
-  const previewPane = document.querySelector('.preview-pane') as HTMLElement
-
-  if (showEditor.value && showPreview.value && customEditorWidth.value !== null && customPreviewWidth.value !== null) {
-    if (editorPane) {
-      editorPane.style.width = `${customEditorWidth.value}%`
-    }
-    if (previewPane) {
-      previewPane.style.width = `${customPreviewWidth.value}%`
-    }
-  } else {
-    if (editorPane) {
-      editorPane.style.width = ''
-    }
-    if (previewPane) {
-      previewPane.style.width = ''
-    }
-  }
-}
-
-const startResize = (e: MouseEvent): void => {
-  e.preventDefault()
-
-  const startX = e.clientX
-  const editorPane = document.querySelector('.editor-pane') as HTMLElement
-  const previewPane = document.querySelector('.preview-pane') as HTMLElement
-
-  if (!editorPane || !previewPane) return
-
-  const startEditorWidth = editorPane.offsetWidth
-  const containerWidth = editorPane.parentElement!.offsetWidth
-
-  const handleMouseMove = (e: MouseEvent): void => {
-    const deltaX = e.clientX - startX
-    const newEditorWidth = startEditorWidth + deltaX
-    const newEditorPercent = (newEditorWidth / containerWidth) * 100
-
-    if (newEditorPercent >= 20 && newEditorPercent <= 80) {
-      editorPane.style.width = `${newEditorPercent}%`
-      previewPane.style.width = `${100 - newEditorPercent}%`
-    }
-  }
-
-  const handleMouseUp = (): void => {
-    document.removeEventListener('mousemove', handleMouseMove)
-    document.removeEventListener('mouseup', handleMouseUp)
-
-    const editorPercent = (editorPane.offsetWidth / containerWidth) * 100
-    const previewPercent = 100 - editorPercent
-
-    customEditorWidth.value = editorPercent
-    customPreviewWidth.value = previewPercent
-    savePaneWidths()
-  }
-
-  document.addEventListener('mousemove', handleMouseMove)
-  document.addEventListener('mouseup', handleMouseUp)
-}
-
-const setupScrollSync = (): void => {
-  if (!editorViewRef.value || !previewContainer.value) return
-
-  let isEditorScrolling = false
-  let isPreviewScrolling = false
-
-  const editorScrollHandler = (): void => {
-    if (isPreviewScrolling) return
-
-    isEditorScrolling = true
-
-    const editorElement = editorViewRef.value!.scrollDOM
-    const scrollPercentage =
-      editorElement.scrollTop / (editorElement.scrollHeight - editorElement.clientHeight)
-
-    const previewElement = previewContainer.value!
-    const maxScrollTop = previewElement.scrollHeight - previewElement.clientHeight
-
-    previewElement.scrollTop = scrollPercentage * maxScrollTop
-
-    setTimeout(() => {
-      isEditorScrolling = false
-    }, 50)
-  }
-
-  const previewScrollHandler = (): void => {
-    if (isEditorScrolling) return
-
-    isPreviewScrolling = true
-
-    const previewElement = previewContainer.value!
-    const scrollPercentage =
-      previewElement.scrollTop / (previewElement.scrollHeight - previewElement.clientHeight)
-
-    const editorElement = editorViewRef.value!.scrollDOM
-    const maxScrollTop = editorElement.scrollHeight - editorElement.clientHeight
-
-    editorElement.scrollTop = scrollPercentage * maxScrollTop
-
-    setTimeout(() => {
-      isPreviewScrolling = false
-    }, 50)
-  }
-
-  editorViewRef.value.scrollDOM.addEventListener('scroll', editorScrollHandler)
-  previewContainer.value.addEventListener('scroll', previewScrollHandler)
-}
-
-const loadSettings = (): void => {
-  try {
-    const savedAutosave = localStorage.getItem('mermditor-autosave')
-    if (savedAutosave !== null) {
-      autosave.value = savedAutosave === 'true'
-    }
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Error loading settings:', error)
-  }
-}
-
-const loadContent = (): void => {
-  const saved = localStorage.getItem('mermditor-content')
-  if (saved) {
-    content.value = saved
-  }
-}
-
-const savePaneWidths = (): void => {
-  try {
-    if (customEditorWidth.value !== null && customPreviewWidth.value !== null) {
-      localStorage.setItem('mermditor-editor-width', customEditorWidth.value.toString())
-      localStorage.setItem('mermditor-preview-width', customPreviewWidth.value.toString())
-    }
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Error saving pane widths:', error)
-  }
-}
-
-const loadPaneWidths = (): void => {
-  try {
-    const savedEditorWidth = localStorage.getItem('mermditor-editor-width')
-    const savedPreviewWidth = localStorage.getItem('mermditor-preview-width')
-
-    if (savedEditorWidth && savedPreviewWidth) {
-      customEditorWidth.value = parseFloat(savedEditorWidth)
-      customPreviewWidth.value = parseFloat(savedPreviewWidth)
-    }
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Error loading pane widths:', error)
-  }
-}
-
-// Watchers
-let saveTimeout: ReturnType<typeof setTimeout> | null = null
-
-watch(content, () => {
-  if (autosave.value) {
-    if (saveTimeout) {
-      clearTimeout(saveTimeout)
-    }
-    saveTimeout = setTimeout(() => {
+  watch(autosave, (newValue) => {
+    if (newValue) {
       actions.saveContent()
-    }, 10000)
-  }
-})
+    }
+  })
 
-watch(autosave, (newValue) => {
-  if (newValue) {
+  const onAutosaveToggle = (checked: boolean): void => {
+    pendingAutosaveValue.value = checked
+    if (checked) {
+      confirmAutosaveOn.value = true
+    } else {
+      confirmAutosaveOff.value = true
+    }
+  }
+
+  const cancelAutosaveChange = (): void => {
+    confirmAutosaveOn.value = false
+    confirmAutosaveOff.value = false
+    pendingAutosaveValue.value = null
+  }
+
+  const confirmEnableAutosave = (): void => {
+    autosave.value = true
+    confirmAutosaveOn.value = false
+    pendingAutosaveValue.value = null
+    try {
+      localStorage.setItem('mermditor-autosave', 'true')
+    } catch {
+      /* ignore storage errors */
+    }
     actions.saveContent()
   }
-})
 
-const onAutosaveToggle = (checked: boolean): void => {
-  pendingAutosaveValue.value = checked
-  if (checked) {
-    confirmAutosaveOn.value = true
-  } else {
-    confirmAutosaveOff.value = true
-  }
-}
-
-const cancelAutosaveChange = (): void => {
-  confirmAutosaveOn.value = false
-  confirmAutosaveOff.value = false
-  pendingAutosaveValue.value = null
-}
-
-const confirmEnableAutosave = (): void => {
-  autosave.value = true
-  confirmAutosaveOn.value = false
-  pendingAutosaveValue.value = null
-  try {
-    localStorage.setItem('mermditor-autosave', 'true')
-  } catch {
-    /* ignore storage errors */
-  }
-  actions.saveContent()
-}
-
-const confirmDisableAutosave = (): void => {
-  autosave.value = false
-  confirmAutosaveOff.value = false
-  pendingAutosaveValue.value = null
-  try {
-    localStorage.setItem('mermditor-autosave', 'false')
-    localStorage.removeItem('mermditor-content')
-  } catch {
-    /* ignore storage errors */
-  }
-}
-
-const onClearStorageClick = (): void => {
-  confirmClearData.value = true
-}
-
-const confirmClearDataNow = (): void => {
-  try {
-    localStorage.removeItem('mermditor-content')
-    localStorage.removeItem('mermditor-autosave')
-    localStorage.removeItem('mermditor-editor-width')
-    localStorage.removeItem('mermditor-preview-width')
-    localStorage.removeItem('mermditor-recent-emojis')
+  const confirmDisableAutosave = (): void => {
     autosave.value = false
-  } catch {
-    /* ignore storage errors */
+    confirmAutosaveOff.value = false
+    pendingAutosaveValue.value = null
+    try {
+      localStorage.setItem('mermditor-autosave', 'false')
+      localStorage.removeItem('mermditor-content')
+    } catch {
+      /* ignore storage errors */
+    }
   }
-  confirmClearData.value = false
-}
 
-// Check for mobile device
-const checkMobile = (): void => {
-  isMobile.value = window.innerWidth < 640
-}
+  const onClearStorageClick = (): void => {
+    confirmClearData.value = true
+  }
 
-// Lifecycle
-onMounted(async () => {
-  checkMobile()
-  window.addEventListener('resize', checkMobile)
+  const confirmClearDataNow = (): void => {
+    try {
+      localStorage.removeItem('mermditor-content')
+      localStorage.removeItem('mermditor-autosave')
+      localStorage.removeItem('mermditor-editor-width')
+      localStorage.removeItem('mermditor-preview-width')
+      localStorage.removeItem('mermditor-recent-emojis')
+      autosave.value = false
+    } catch {
+      /* ignore storage errors */
+    }
+    confirmClearData.value = false
+  }
 
-  loadContent()
-  loadSettings()
-  loadPaneWidths()
-  await nextTick()
-  initEditor()
+  // Check for mobile device
+  const checkMobile = (): void => {
+    isMobile.value = window.innerWidth < 640
+  }
 
-  setTimeout(() => {
-    setupScrollSync()
-    applyPaneWidths()
+  // Lifecycle
+  onMounted(async () => {
+    checkMobile()
+    window.addEventListener('resize', checkMobile)
+
+    loadContent()
+    loadSettings()
+    loadPaneWidths()
+    await nextTick()
+    initEditor()
 
     setTimeout(() => {
-      isLoading.value = false
-    }, 500) // Small delay to ensure smooth transition
-  }, 500)
-})
+      setupScrollSync()
+      applyPaneWidths()
 
-// Cleanup
-onUnmounted(() => {
-  window.removeEventListener('resize', checkMobile)
-})
+      setTimeout(() => {
+        isLoading.value = false
+      }, 500) // Small delay to ensure smooth transition
+    }, 500)
+  })
+
+  // Cleanup
+  onUnmounted(() => {
+    window.removeEventListener('resize', checkMobile)
+  })
 </script>
