+++ package.json
@@ -27,14 +27,21 @@
     "@nuxtjs/tailwindcss": "^6.14.0",
     "@phosphor-icons/vue": "^2.2.1",
     "@tailwindcss/typography": "^0.5.18",
+    "@types/markdown-it": "^14.1.2",
     "@types/node": "^24.5.2",
     "@vercel/analytics": "^1.5.0",
     "@vercel/speed-insights": "^1.2.0",
     "codemirror": "^6.0.2",
     "eslint": "^9.36.0",
     "katex": "^0.16.22",
-    "marked": "^16.3.0",
-    "marked-emoji": "^2.0.1",
+    "markdown-it": "^14.1.0",
+    "markdown-it-deflist": "^3.0.0",
+    "markdown-it-emoji": "^3.0.0",
+    "markdown-it-footnote": "^4.0.0",
+    "markdown-it-mark": "^4.0.0",
+    "markdown-it-sub": "^2.0.0",
+    "markdown-it-sup": "^2.0.0",
+    "markdown-it-task-lists": "^2.1.1",
     "mermaid": "^11.12.0",
     "nuxt": "^4.1.2",
     "prettier": "^3.6.2",
